Moved `Base.Heliotonic` to `Base.Scale.Heliotonic` and `Base.Interval` to `Base.Core.Interval`
